Added test for generating ecmwf warnings
