+++ tests/compare/rapid_output/dominican_republic-haina/20170708.00/return_20_points.geojson
@@ -1,0 +1,10 @@
+{
+ "crs": {
+  "type": "name",
+  "properties": {
+   "name": "EPSG:4326"
+  }
+ },
+ "type": "FeatureCollection",
+ "features": []
+}
