+++ tests/compare/rapid_output/dominican_republic-haina/20170708.00/return_2_points.geojson
@@ -1,0 +1,235 @@
+{
+ "crs": {
+  "type": "name",
+  "properties": {
+   "name": "EPSG:4326"
+  }
+ },
+ "type": "FeatureCollection",
+ "features": [
+  {
+   "geometry": {
+    "type": "Point",
+    "coordinates": [
+     -70.17987023499995,
+     18.59684929900004
+    ]
+   },
+   "type": "Feature",
+   "properties": {
+    "peak_date": "2017-07-09",
+    "size": 0,
+    "mean_plus_std_peak": 12.37,
+    "rivid": 21893
+   }
+  },
+  {
+   "geometry": {
+    "type": "Point",
+    "coordinates": [
+     -70.17987023499995,
+     18.59684929900004
+    ]
+   },
+   "type": "Feature",
+   "properties": {
+    "peak_date": "2017-07-10",
+    "size": 0,
+    "mean_plus_std_peak": 13.19,
+    "rivid": 21893
+   }
+  },
+  {
+   "geometry": {
+    "type": "Point",
+    "coordinates": [
+     -70.18478411299998,
+     18.66302281000003
+    ]
+   },
+   "type": "Feature",
+   "properties": {
+    "peak_date": "2017-07-09",
+    "mean_peak": 10.91,
+    "size": 1,
+    "rivid": 21890
+   }
+  },
+  {
+   "geometry": {
+    "type": "Point",
+    "coordinates": [
+     -70.18478411299998,
+     18.66302281000003
+    ]
+   },
+   "type": "Feature",
+   "properties": {
+    "peak_date": "2017-07-09",
+    "size": 0,
+    "mean_plus_std_peak": 23.54,
+    "rivid": 21890
+   }
+  },
+  {
+   "geometry": {
+    "type": "Point",
+    "coordinates": [
+     -70.18478411299998,
+     18.66302281000003
+    ]
+   },
+   "type": "Feature",
+   "properties": {
+    "peak_date": "2017-07-10",
+    "mean_peak": 11.2,
+    "size": 1,
+    "rivid": 21890
+   }
+  },
+  {
+   "geometry": {
+    "type": "Point",
+    "coordinates": [
+     -70.18478411299998,
+     18.66302281000003
+    ]
+   },
+   "type": "Feature",
+   "properties": {
+    "peak_date": "2017-07-10",
+    "size": 0,
+    "mean_plus_std_peak": 20.37,
+    "rivid": 21890
+   }
+  },
+  {
+   "geometry": {
+    "type": "Point",
+    "coordinates": [
+     -70.18478411299998,
+     18.66302281000003
+    ]
+   },
+   "type": "Feature",
+   "properties": {
+    "peak_date": "2017-07-14",
+    "size": 0,
+    "mean_plus_std_peak": 11.59,
+    "rivid": 21890
+   }
+  },
+  {
+   "geometry": {
+    "type": "Point",
+    "coordinates": [
+     -70.18478411299998,
+     18.66302281000003
+    ]
+   },
+   "type": "Feature",
+   "properties": {
+    "peak_date": "2017-07-16",
+    "size": 0,
+    "mean_plus_std_peak": 12.68,
+    "rivid": 21890
+   }
+  },
+  {
+   "geometry": {
+    "type": "Point",
+    "coordinates": [
+     -70.23344103099998,
+     18.72881291300007
+    ]
+   },
+   "type": "Feature",
+   "properties": {
+    "peak_date": "2017-07-08",
+    "size": 0,
+    "mean_plus_std_peak": 11.12,
+    "rivid": 21841
+   }
+  },
+  {
+   "geometry": {
+    "type": "Point",
+    "coordinates": [
+     -70.08183596899994,
+     18.510065534000034
+    ]
+   },
+   "type": "Feature",
+   "properties": {
+    "peak_date": "2017-07-09",
+    "size": 0,
+    "mean_plus_std_peak": 16.83,
+    "rivid": 22074
+   }
+  },
+  {
+   "geometry": {
+    "type": "Point",
+    "coordinates": [
+     -70.08183596899994,
+     18.510065534000034
+    ]
+   },
+   "type": "Feature",
+   "properties": {
+    "peak_date": "2017-07-10",
+    "mean_peak": 12.17,
+    "size": 1,
+    "rivid": 22074
+   }
+  },
+  {
+   "geometry": {
+    "type": "Point",
+    "coordinates": [
+     -70.08183596899994,
+     18.510065534000034
+    ]
+   },
+   "type": "Feature",
+   "properties": {
+    "peak_date": "2017-07-10",
+    "size": 0,
+    "mean_plus_std_peak": 19.96,
+    "rivid": 22074
+   }
+  },
+  {
+   "geometry": {
+    "type": "Point",
+    "coordinates": [
+     -70.08183596899994,
+     18.510065534000034
+    ]
+   },
+   "type": "Feature",
+   "properties": {
+    "peak_date": "2017-07-14",
+    "size": 0,
+    "mean_plus_std_peak": 11.72,
+    "rivid": 22074
+   }
+  },
+  {
+   "geometry": {
+    "type": "Point",
+    "coordinates": [
+     -70.08183596899994,
+     18.510065534000034
+    ]
+   },
+   "type": "Feature",
+   "properties": {
+    "peak_date": "2017-07-16",
+    "size": 0,
+    "mean_plus_std_peak": 12.8,
+    "rivid": 22074
+   }
+  }
+ ]
+}
