--- tests/input/historical_input/dominican_republic-haina/.gitignore
@@ -1,2 +1,0 @@
-*
-!.gitignore
